length for line segment to CGFloat

+++ GridOptic/GridOptic/GridOptic/OpticRepresentation/GOFlatLensRep.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 class GOFlatLensRep: GOFlatOpticRep {
-    override init(center: GOCoordinate, thickness: NSInteger, length: NSInteger, direction: CGVector, refractionIndex: CGFloat, id: String) {
+    override init(center: GOCoordinate, thickness: CGFloat, length: CGFloat, direction: CGVector, refractionIndex: CGFloat, id: String) {
         super.init(center: center, thickness: thickness, length: length, direction: direction, refractionIndex: refractionIndex, id: id)
         self.setDeviceType(DeviceType.Lens)
     }
