+++ GridOptic/GridOptic/GridOptic/OpticRepresentation/GOFlatWallRep.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 class GOFlatWallRep: GOFlatOpticRep {
-    override init(center: GOCoordinate, thickness: NSInteger, length: NSInteger, direction: CGVector, id: String) {
+    override init(center: GOCoordinate, thickness: CGFloat, length: CGFloat, direction: CGVector, id: String) {
         super.init(center: center, thickness: thickness, length: length, direction: direction, id: id)
         self.setDeviceType(DeviceType.Wall)
     }
